fix: change return type of getLongClickCount from uint8_t to uint16_t for improved count handling

+++ src/Button2.cpp
@@ -455,7 +455,7 @@ void Button2::_checkForLongClick(unsigned long now) {
 
 /////////////////////////////////////////////////////////////////
 
-uint8_t Button2::getLongClickCount() const {
+uint16_t Button2::getLongClickCount() const {
   return longclick_counter;
 }
